Adjusted EditDateRangeButton location; js_only.
