+++ assets/src/application/ui/layout/EntityCard/styles.scss
@@ -23,6 +23,7 @@
 		.ee-edit-calendar-date-range {
 			display: flex;
 			justify-content: center;
+			margin: 0.25rem 0;
 			padding: 0 0 var(--ee-padding-nano);
 		}
 	}
