+++ assets/src/application/ui/layout/EntityCard/types.ts
@@ -7,5 +7,5 @@ export interface EntityCardProps {
 	entity: Entity;
 	reverse?: boolean;
 	sidebar: JSX.Element;
-	sidebarClass?: string;
+	sidebarClassName?: string;
 }
