+++ assets/src/domain/eventEditor/ui/datetimes/datesList/cardView/DateCard.tsx
@@ -1,35 +1,21 @@
-import React, { useCallback } from 'react';
+import React from 'react';
 import { __ } from '@wordpress/i18n';
 
 import DateActionsMenu from '../actionsMenu/DateActionsMenu';
 import Details from './Details';
 import EntityCard from '@appLayout/EntityCard';
-import { CalendarDateSwitcher, EditDateRangeButton } from '@appCalendars/dateDisplay';
+import { CalendarDateSwitcher } from '@appCalendars/dateDisplay';
 import { DatetimeProvider } from '@edtrServices/context/DatetimeContext';
 import { EntityActionsMenuLayout } from '@appLayout/entityActionsMenu';
 import { getPropsAreEqual } from '@appServices/utilities';
 import { getStatusTextLabel, statusBgColorClassName } from '@sharedEntities/datetimes/helpers';
 import { useDatesListFilterState } from '@edtrServices/filterState';
-import { useDatetimeMutator } from '@edtrServices/apollo/mutations';
 import { useMemoStringify } from '@application/services/hooks';
-import { useTimeZoneTime } from '@appServices/hooks';
 import type { DateItemProps } from '../types';
 
 const DateCard: React.FC<DateItemProps> = ({ entity: date }) => {
 	const { displayStartOrEndDate } = useDatesListFilterState();
-	const { updateEntity } = useDatetimeMutator(date.id);
-	const { siteTimeToUtc } = useTimeZoneTime();
 
-	const onEditHandler = useCallback(
-		(dates: string[]): void => {
-			const [start, end] = dates;
-			// convert start & end dates to proper UTC "startDate" and "endDate"
-			const startDate = siteTimeToUtc(new Date(start)).toISOString();
-			const endDate = siteTimeToUtc(new Date(end)).toISOString();
-			updateEntity({ startDate, endDate });
-		},
-		[date.cacheId, updateEntity]
-	);
 	const bgClassName = statusBgColorClassName(date);
 	const footer = getStatusTextLabel(date);
 	const labels = useMemoStringify({ footer });
@@ -40,13 +26,6 @@ const DateCard: React.FC<DateItemProps> = ({ entity: date }) => {
 				endDate={date.endDate}
 				labels={labels}
 				startDate={date.startDate}
-			/>
-			<EditDateRangeButton
-				endDate={date.endDate}
-				header={__('Edit Event Date Start and End Dates')}
-				onEditHandler={onEditHandler}
-				startDate={date.startDate}
-				tooltip={__("click to edit this event date's start and end dates")}
 			/>
 		</>
 	);
@@ -59,7 +38,7 @@ const DateCard: React.FC<DateItemProps> = ({ entity: date }) => {
 				details={<Details entity={date} />}
 				entity={date}
 				sidebar={sidebar}
-				sidebarClass={bgClassName}
+				sidebarClassName={bgClassName}
 			/>
 		</DatetimeProvider>
 	) : null;
